Former-commit-id: 84215636e691e7677bc545eed674e32f6c553015 [formerly 938712a92ef0fed66f1bf6d75da9d42d06067e08] [formerly 57abc0ebb912cdf3b3b90c75a630d1956f947fbb [formerly 1e7651d7769c619f900f90c9b83caf71d95a03c5]]
Former-commit-id: 57abc0ebb912cdf3b3b90c75a630d1956f947fbb
Former-commit-id: 7fe733ddfc6c2d9093f61212483b4cb1936afa3b

+++ front_end/src/components/StockItemDetail/BoardDetail.jsx
@@ -48,8 +48,10 @@ export default function BoardDetail() {
 
   // 게시글 삭제
   const clickDeleteBoard = async () => {
-    await deleteBoard(boardId);
-    navigate(`/stock/${id}/board`);
+    if (window.confirm("게시글을 삭제하시겠습니까?")) {
+      await deleteBoard(boardId);
+      navigate(`/stock/${id}/board`);
+    }
   };
 
   // 게시글 수정
